feat: move body to common component for both response & request
